feat(gui): replace old images with new mockups and update Slide05 content
