+++ src-gui/src/renderer/components/modal/introduction/slides/Slide02_ChooseAMaker.tsx
@@ -1,6 +1,6 @@
 import { Typography } from '@material-ui/core'
 import SlideTemplate from './SlideTemplate'
-import imagePath from '../../../../../assets/selectAMaker.png'
+import imagePath from '../../../../../assets/mockMakerSelection.webp'
 
 export default function Slide02_ChooseAMaker(props: slideProps) {
     return (
@@ -9,10 +9,9 @@ export default function Slide02_ChooseAMaker(props: slideProps) {
             stepLabel="Step 1"
             {...props}
             imagePath={imagePath}
-            imagePadded
         >
             <Typography variant="subtitle1">
-                Makers provide liquidity in XMR which which can be exchanged for
+                Makers provide liquidity in XMR which can be exchanged for
                 BTC. They have varying Min and Max amounts they can swap and
                 charge varying markup fees. Makers can be found with
             </Typography>
